server/requirements.txt
asgiref==3.5.0
click==8.1.3
colorama==0.4.4
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
importlib-metadata==4.11.3
joblib==1.1.0
nltk==3.7
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
regex==2022.4.24
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
typing_extensions==4.1.1
zipp==3.8.0
